Var resolution for query fns/preds when in JVM (#178)
This commit adds a reader conditional to enable the use of
namespace-qualified symbols as query functions or predicates, when in
JVM Clojure. This allows queries using functions or predicates to be
reused between Datomic and DataScript when running on the JVM.

The semantics of a DataScript query are otherwise unchanged; global var
resolution is only attempted as a final step before failure.

The semantics when operating in a ClojureScript context are entirely
unchanged, since global symbol->var resolution is not supported.